requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.1
xlsxwriter==3.2.5
